Feature/weights

+++ db/migrate/20200715161205_create_diets.rb
@@ -5,7 +5,7 @@ class CreateDiets < ActiveRecord::Migration[6.0]
       t.date :end_date, null: false, default: Date.current
       t.integer :initial_weight, null: false
       t.integer :target_weight, null: false
-      t.integer :height
+      t.float :height
       t.references :user, null: false, index: true, foreign_key: true
 
       t.timestamps
